Removes overrides and moves files.
