+++ src/renderer/App.tsx
@@ -1,9 +1,9 @@
 import { CssBaseline, ThemeProvider } from '@mui/material'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
-import { Home } from './components/Home'
 import { ApiProvider } from './contexts/ApiContext'
 import { IntlProvider } from './contexts/IntlContext'
+import { Home } from './pages/Home'
 import { theme } from './Theme'
 
 const queryClient = new QueryClient()
